+++ src/Api/Controller/TwoFactorUpdateController.php
@@ -49,6 +49,10 @@ class TwoFactorUpdateController implements RequestHandlerInterface
             $actor->twofa_secret = '';
             $actor->twofa_codes = '';
         } else {
+            if (!$actor->can("nearata-twofactor.enable")) {
+                return new EmptyResponse(403);
+            }
+
             $actor->twofa_active = true;
             $actor->twofa_secret = $secret;
         }