+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.1.0] - 2022-08-26
 
 - Add php 8 support ([Issue #7](https://github.com/Nearata/flarum-ext-twofactor/issues/7)).
 - Ability to choose which groups can enable Two Factor.